twitter/users: minor refactor
